# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Tailor API
  version: 0.1.0
paths:
  /api/health:
    get:
      operationId: getHealth
      responses:
        "200":
          description: Server health.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
  /api/topology:
    get:
      operationId: getTopology
      responses:
        "200":
          description: Current tailnet topology snapshot. The topology socket sends the same shape as a topology.snapshot payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TopologyResponse"
        "503":
          description: Tailscale LocalAPI is unavailable.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LocalAPIStatusResponse"
  /api/status:
    get:
      operationId: getStatus
      responses:
        "200":
          description: LocalAPI connectivity status.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LocalAPIStatusResponse"
  /api/cloud/status:
    get:
      operationId: getCloudStatus
      responses:
        "200":
          description: Cloud API authentication status.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CloudAuthStatusResponse"
  /api/cloud/auth:
    post:
      operationId: authenticateCloud
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CloudAuthRequest"
      responses:
        "200":
          description: API key authentication succeeded and policy was fetched.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CloudAuthStatusResponse"
        "400":
          description: Invalid auth request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Tailscale Cloud API request failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/policy:
    get:
      operationId: getPolicy
      responses:
        "200":
          description: Current tailnet policy file as raw HuJSON.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyResponse"
        "401":
          description: Cloud API authentication has not been enabled.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Tailscale Cloud API request failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/policy/map:
    get:
      operationId: getPolicyMap
      responses:
        "200":
          description: Read-only structured view of the current tailnet policy file.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyMapResponse"
        "401":
          description: Cloud API authentication has not been enabled.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Tailscale Cloud API request failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/policy/draft:
    post:
      operationId: draftPolicyAclRule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PolicyDraftRequest"
      responses:
        "200":
          description: Draft policy with appended ACL rule.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyDraftResponse"
        "400":
          description: Invalid draft request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Cloud API authentication has not been enabled.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/policy/evaluate-draft:
    post:
      operationId: evaluatePolicyDraft
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PolicyEvaluateDraftRequest"
      responses:
        "200":
          description: Graph-ready access impact for an unsaved draft policy.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyEvaluateDraftResponse"
        "400":
          description: Invalid draft policy.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Cloud API authentication has not been enabled.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Tailscale LocalAPI is unavailable.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LocalAPIStatusResponse"
  /api/policy/validate:
    post:
      operationId: validatePolicyDraft
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PolicyValidateRequest"
      responses:
        "200":
          description: Draft policy is valid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyValidateResponse"
        "502":
          description: Draft policy failed validation or Cloud API request failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyValidateResponse"
  /api/policy/save:
    post:
      operationId: saveValidatedPolicyDraft
      responses:
        "200":
          description: Last validated draft was saved.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicySaveResponse"
        "401":
          description: Cloud API authentication has not been enabled.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Save failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthResponse:
      type: object
      required: [status, version]
      properties:
        status:
          type: string
        version:
          type: string
    LocalAPIStatusResponse:
      type: object
      required: [available, localApiEndpoint]
      properties:
        available:
          type: boolean
        localApiEndpoint:
          type: string
        error:
          type: string
    Device:
      type: object
      required: [id, name, ip, tailscaleIps, os, online, owner, tags, subnetRouter, routedSubnets]
      properties:
        id:
          type: string
        name:
          type: string
        ip:
          type: string
        tailscaleIps:
          type: array
          items:
            type: string
        os:
          type: string
        online:
          type: boolean
        owner:
          type: string
        tags:
          type: array
          items:
            type: string
        subnetRouter:
          type: boolean
        routedSubnets:
          type: array
          items:
            type: string
        lastSeen:
          type: string
    EdgeKind:
      type: string
      enum: [owner, tag, subnet, acl]
    Edge:
      type: object
      required: [id, from, to, kind]
      properties:
        id:
          type: string
        from:
          type: string
        to:
          type: string
        kind:
          $ref: "#/components/schemas/EdgeKind"
        labels:
          type: array
          items:
            type: string
        protocols:
          type: array
          items:
            type: string
        ports:
          type: array
          items:
            type: string
        accessScope:
          $ref: "#/components/schemas/AccessScope"
        policyRefs:
          type: array
          items:
            $ref: "#/components/schemas/PolicyRef"
        perspectives:
          type: array
          items:
            type: string
    AccessScope:
      type: string
      enum: [ssh, http, broad, custom, limited, none, ""]
    PolicyRef:
      type: object
      required: [section, index]
      properties:
        section:
          type: string
        index:
          type: integer
        src:
          type: string
        dst:
          type: string
    TopologyResponse:
      type: object
      required: [devices, edges, tailnet]
      properties:
        devices:
          type: array
          items:
            $ref: "#/components/schemas/Device"
        edges:
          type: array
          items:
            $ref: "#/components/schemas/Edge"
        tailnet:
          type: string
    CloudAuthRequest:
      type: object
      required: [tailnet, apiKey]
      properties:
        tailnet:
          type: string
        apiKey:
          type: string
    CloudAuthStatusResponse:
      type: object
      required: [authenticated, hasPolicy]
      properties:
        authenticated:
          type: boolean
        tailnet:
          type: string
        hasPolicy:
          type: boolean
        devMode:
          type: boolean
          description: True when authenticated with the in-memory demo tailnet key.
    PolicyResponse:
      type: object
      required: [tailnet, hujson]
      properties:
        tailnet:
          type: string
        hujson:
          type: string
    PolicyMapResponse:
      type: object
      required: [tailnet, hujson, sections]
      properties:
        tailnet:
          type: string
        hujson:
          type: string
        sections:
          type: array
          items:
            $ref: "#/components/schemas/PolicySection"
        parseError:
          type: string
    PolicySection:
      type: object
      required: [name, type, supported, count]
      properties:
        name:
          type: string
        type:
          type: string
        supported:
          type: boolean
        count:
          type: integer
        entries:
          type: array
          items:
            $ref: "#/components/schemas/PolicySectionEntry"
        raw:
          description: Original decoded section value.
          nullable: true
        description:
          type: string
    PolicySectionEntry:
      type: object
      required: [label]
      properties:
        label:
          type: string
        summary:
          type: string
        selectors:
          type: array
          items:
            type: string
        value:
          description: Original decoded entry value.
          nullable: true
    PolicyDraftRequest:
      type: object
      required: [sources, destinations, ports]
      properties:
        sources:
          type: array
          items:
            type: string
        destinations:
          type: array
          items:
            type: string
        ports:
          type: array
          items:
            type: string
        protocol:
          type: string
    ACLDraft:
      type: object
      required: [action, src, dst]
      properties:
        action:
          type: string
        src:
          type: array
          items:
            type: string
        dst:
          type: array
          items:
            type: string
        proto:
          type: string
    PolicyDraftResponse:
      type: object
      required: [tailnet, rule, hujson]
      properties:
        tailnet:
          type: string
        rule:
          $ref: "#/components/schemas/ACLDraft"
        hujson:
          type: string
    PolicyEvaluateDraftRequest:
      type: object
      required: [hujson]
      properties:
        hujson:
          type: string
        perspective:
          type: string
    PolicyEdgeChange:
      type: object
      required: [state, edge]
      properties:
        state:
          type: string
          enum: [added, removed, unchanged, changed]
        edge:
          $ref: "#/components/schemas/Edge"
        saved:
          $ref: "#/components/schemas/Edge"
        draft:
          $ref: "#/components/schemas/Edge"
    UnresolvedSelector:
      type: object
      required: [section, index, selector, role]
      properties:
        section:
          type: string
        index:
          type: integer
        selector:
          type: string
        role:
          type: string
    ApplicationGrant:
      type: object
      required: [section, index, src, dst, capabilities]
      properties:
        section:
          type: string
        index:
          type: integer
        src:
          type: array
          items:
            type: string
        dst:
          type: array
          items:
            type: string
        capabilities:
          type: array
          items:
            type: string
    PolicyEvaluateDraftResponse:
      type: object
      required:
        [
          tailnet,
          added,
          removed,
          unchanged,
          changed,
          broadAccess,
          visibleDeviceIds,
          unresolvedSelectors,
          unsupportedSections,
          applicationGrants,
        ]
      properties:
        tailnet:
          type: string
        added:
          type: array
          items:
            $ref: "#/components/schemas/PolicyEdgeChange"
        removed:
          type: array
          items:
            $ref: "#/components/schemas/PolicyEdgeChange"
        unchanged:
          type: array
          items:
            $ref: "#/components/schemas/PolicyEdgeChange"
        changed:
          type: array
          items:
            $ref: "#/components/schemas/PolicyEdgeChange"
        broadAccess:
          type: array
          items:
            $ref: "#/components/schemas/Edge"
        visibleDeviceIds:
          type: array
          items:
            type: string
          description: >
            Device IDs in the perspective trimmed netmap (Tailscale device visibility).
            When no perspective is supplied, all tailnet devices are listed.
        unresolvedSelectors:
          type: array
          items:
            $ref: "#/components/schemas/UnresolvedSelector"
        unsupportedSections:
          type: array
          items:
            type: string
        applicationGrants:
          type: array
          items:
            $ref: "#/components/schemas/ApplicationGrant"
    PolicyValidateRequest:
      type: object
      required: [hujson]
      properties:
        hujson:
          type: string
    PolicyValidateResponse:
      type: object
      required: [valid, tailnet]
      properties:
        valid:
          type: boolean
        tailnet:
          type: string
        errors:
          type: array
          items:
            type: string
    PolicySaveResponse:
      type: object
      required: [saved, tailnet, hujson]
      properties:
        saved:
          type: boolean
        tailnet:
          type: string
        hujson:
          type: string
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string